LPS-69934 Decouple subscription logic from message boards portal-impl to subscription service module using service wrappers
