+++ modules/apps/collaboration/subscription/subscription-service/src/main/java/com/liferay/subscription/internal/service/SubscriptionMBThreadLocalServiceWrapper.java
@@ -1,0 +1,62 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.subscription.internal.service;
+
+import com.liferay.message.boards.kernel.model.MBThread;
+import com.liferay.message.boards.kernel.service.MBThreadLocalService;
+import com.liferay.message.boards.kernel.service.MBThreadLocalServiceWrapper;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.service.ServiceWrapper;
+import com.liferay.portal.kernel.service.SubscriptionLocalService;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Adolfo Pérez
+ */
+@Component(immediate = true, service = ServiceWrapper.class)
+public class SubscriptionMBThreadLocalServiceWrapper
+	extends MBThreadLocalServiceWrapper {
+
+	public SubscriptionMBThreadLocalServiceWrapper() {
+		super(null);
+	}
+
+	public SubscriptionMBThreadLocalServiceWrapper(
+		MBThreadLocalService mbThreadLocalService) {
+
+		super(mbThreadLocalService);
+	}
+
+	@Override
+	public void deleteThread(MBThread thread) throws PortalException {
+		super.deleteThread(thread);
+
+		_subscriptionLocalService.deleteSubscriptions(
+			thread.getCompanyId(), MBThread.class.getName(),
+			thread.getThreadId());
+	}
+
+	@Reference(unbind = "-")
+	protected void setSubscriptionLocalService(
+		SubscriptionLocalService subscriptionLocalService) {
+
+		_subscriptionLocalService = subscriptionLocalService;
+	}
+
+	private SubscriptionLocalService _subscriptionLocalService;
+
+}
