+++ portal-impl/src/com/liferay/portlet/messageboards/service/impl/MBDiscussionLocalServiceImpl.java
@@ -112,17 +112,12 @@ public class MBDiscussionLocalServiceImpl
 	public void subscribeDiscussion(
 			long userId, long groupId, String className, long classPK)
 		throws PortalException {
-
-		subscriptionLocalService.addSubscription(
-			userId, groupId, className, classPK);
 	}
 
 	@Override
 	public void unsubscribeDiscussion(
 			long userId, String className, long classPK)
 		throws PortalException {
-
-		subscriptionLocalService.deleteSubscription(userId, className, classPK);
 	}
 
 }
