+++ portal-impl/src/com/liferay/portlet/messageboards/service/impl/MBMessageLocalServiceImpl.java
@@ -33,23 +33,17 @@ import com.liferay.message.boards.kernel.model.MBThread;
 import com.liferay.message.boards.kernel.model.MBThreadConstants;
 import com.liferay.message.boards.kernel.util.comparator.MessageCreateDateComparator;
 import com.liferay.message.boards.kernel.util.comparator.MessageThreadComparator;
-import com.liferay.portal.kernel.comment.Comment;
 import com.liferay.portal.kernel.dao.orm.QueryDefinition;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
-import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.Company;
-import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.ModelHintsUtil;
 import com.liferay.portal.kernel.model.ResourceConstants;
 import com.liferay.portal.kernel.model.SystemEventConstants;
 import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.notifications.UserNotificationDefinition;
-import com.liferay.portal.kernel.parsers.bbcode.BBCodeTranslatorUtil;
 import com.liferay.portal.kernel.portlet.PortletProvider;
 import com.liferay.portal.kernel.portlet.PortletProviderUtil;
 import com.liferay.portal.kernel.portletfilerepository.PortletFileRepositoryUtil;
@@ -68,41 +62,32 @@ import com.liferay.portal.kernel.service.permission.ModelPermissions;
 import com.liferay.portal.kernel.settings.LocalizedValuesMap;
 import com.liferay.portal.kernel.social.SocialActivityManagerUtil;
 import com.liferay.portal.kernel.systemevent.SystemEvent;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.HtmlUtil;
-import com.liferay.portal.kernel.util.HttpUtil;
 import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.LocalizationUtil;
 import com.liferay.portal.kernel.util.ObjectValuePair;
 import com.liferay.portal.kernel.util.OrderByComparator;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.PortalUtil;
-import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.SubscriptionSender;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.kernel.workflow.WorkflowHandlerRegistryUtil;
 import com.liferay.portal.kernel.workflow.WorkflowThreadLocal;
 import com.liferay.portal.linkback.LinkbackProducerUtil;
 import com.liferay.portal.util.LayoutURLUtil;
-import com.liferay.portal.util.PrefsPropsUtil;
 import com.liferay.portal.util.PropsValues;
 import com.liferay.portlet.messageboards.MBGroupServiceSettings;
 import com.liferay.portlet.messageboards.model.impl.MBCategoryImpl;
 import com.liferay.portlet.messageboards.model.impl.MBMessageDisplayImpl;
 import com.liferay.portlet.messageboards.service.base.MBMessageLocalServiceBaseImpl;
-import com.liferay.portlet.messageboards.service.permission.MBPermission;
 import com.liferay.portlet.messageboards.social.MBActivityKeys;
 import com.liferay.portlet.messageboards.util.MBSubscriptionSender;
 import com.liferay.portlet.messageboards.util.MBUtil;
-import com.liferay.portlet.messageboards.util.MailingListThreadLocal;
 import com.liferay.social.kernel.model.SocialActivityConstants;
 import com.liferay.trash.kernel.util.TrashUtil;
 
@@ -118,7 +103,6 @@ import java.util.Comparator;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 
 import javax.portlet.PortletRequest;
@@ -660,12 +644,6 @@ public class MBMessageLocalServiceImpl extends MBMessageLocalServiceBaseImpl {
 				PortletFileRepositoryUtil.deletePortletFolder(
 					threadAttachmentsFolderId);
 			}
-
-			// Subscriptions
-
-			subscriptionLocalService.deleteSubscriptions(
-				message.getCompanyId(), MBThread.class.getName(),
-				message.getThreadId());
 
 			// Thread
 
@@ -1510,22 +1488,11 @@ public class MBMessageLocalServiceImpl extends MBMessageLocalServiceBaseImpl {
 	@Override
 	public void subscribeMessage(long userId, long messageId)
 		throws PortalException {
-
-		MBMessage message = mbMessagePersistence.findByPrimaryKey(messageId);
-
-		subscriptionLocalService.addSubscription(
-			userId, message.getGroupId(), MBThread.class.getName(),
-			message.getThreadId());
 	}
 
 	@Override
 	public void unsubscribeMessage(long userId, long messageId)
 		throws PortalException {
-
-		MBMessage message = mbMessagePersistence.findByPrimaryKey(messageId);
-
-		subscriptionLocalService.deleteSubscription(
-			userId, MBThread.class.getName(), message.getThreadId());
 	}
 
 	@Override
@@ -1934,13 +1901,6 @@ public class MBMessageLocalServiceImpl extends MBMessageLocalServiceBaseImpl {
 				}
 			}
 
-			// Subscriptions
-
-			notifySubscribers(
-				userId, (MBMessage)message.clone(),
-				(String)workflowContext.get(WorkflowConstants.CONTEXT_URL),
-				serviceContext);
-
 			// Indexer
 
 			indexer.reindex(message);
@@ -2068,6 +2028,10 @@ public class MBMessageLocalServiceImpl extends MBMessageLocalServiceBaseImpl {
 		return subject;
 	}
 
+	/**
+	 * @deprecated As of 7.0.0, with no direct replacement.
+	 */
+	@Deprecated
 	protected MBSubscriptionSender getSubscriptionSender(
 		long userId, MBCategory category, MBMessage message, String messageURL,
 		String entryTitle, boolean htmlFormat, String messageBody,
@@ -2078,357 +2042,26 @@ public class MBMessageLocalServiceImpl extends MBMessageLocalServiceBaseImpl {
 		LocalizedValuesMap bodyLocalizedValuesMap,
 		ServiceContext serviceContext) {
 
-		MBSubscriptionSender subscriptionSender = new MBSubscriptionSender(
-			MBPermission.RESOURCE_NAME);
-
-		subscriptionSender.setAnonymous(message.isAnonymous());
-		subscriptionSender.setBulk(PropsValues.MESSAGE_BOARDS_EMAIL_BULK);
-		subscriptionSender.setClassName(message.getModelClassName());
-		subscriptionSender.setClassPK(message.getMessageId());
-		subscriptionSender.setCompanyId(message.getCompanyId());
-		subscriptionSender.setContextAttribute(
-			"[$MESSAGE_BODY$]", messageBody, false);
-
-		long groupId = message.getGroupId();
-
-		if (category.getCategoryId() !=
-				MBCategoryConstants.DEFAULT_PARENT_CATEGORY_ID) {
-
-			subscriptionSender.setContextAttribute(
-				"[$CATEGORY_NAME$]", category.getName(), true);
-		}
-		else {
-			subscriptionSender.setLocalizedContextAttributeWithFunction(
-				"[$CATEGORY_NAME$]",
-				locale -> _getLocalizedRootCategoryName(groupId, locale));
-		}
-
-		subscriptionSender.setContextAttributes(
-			"[$MAILING_LIST_ADDRESS$]", replyToAddress, "[$MESSAGE_ID$]",
-			message.getMessageId(), "[$MESSAGE_SUBJECT$]", messageSubject,
-			"[$MESSAGE_SUBJECT_PREFIX$]", messageSubjectPrefix,
-			"[$MESSAGE_URL$]", messageURL, "[$MESSAGE_USER_ADDRESS$]",
-			emailAddress, "[$MESSAGE_USER_NAME$]", fullName);
-		subscriptionSender.setCurrentUserId(userId);
-		subscriptionSender.setEntryTitle(entryTitle);
-		subscriptionSender.setEntryURL(messageURL);
-		subscriptionSender.setFrom(fromAddress, fromName);
-		subscriptionSender.setFullName(fullName);
-		subscriptionSender.setHtmlFormat(htmlFormat);
-		subscriptionSender.setInReplyTo(inReplyTo);
-
-		if (bodyLocalizedValuesMap != null) {
-			subscriptionSender.setLocalizedBodyMap(
-				LocalizationUtil.getMap(bodyLocalizedValuesMap));
-		}
-
-		if (subjectLocalizedValuesMap != null) {
-			subscriptionSender.setLocalizedSubjectMap(
-				LocalizationUtil.getMap(subjectLocalizedValuesMap));
-		}
-
-		Date modifiedDate = message.getModifiedDate();
-
-		subscriptionSender.setMailId(
-			MBUtil.MESSAGE_POP_PORTLET_PREFIX, message.getCategoryId(),
-			message.getMessageId(), modifiedDate.getTime());
-
-		int notificationType =
-			UserNotificationDefinition.NOTIFICATION_TYPE_ADD_ENTRY;
-
-		if (serviceContext.isCommandUpdate()) {
-			notificationType =
-				UserNotificationDefinition.NOTIFICATION_TYPE_UPDATE_ENTRY;
-		}
-
-		subscriptionSender.setNotificationType(notificationType);
-
-		String portletId = PortletProviderUtil.getPortletId(
-			MBMessage.class.getName(), PortletProvider.Action.VIEW);
-
-		subscriptionSender.setPortletId(portletId);
-
-		subscriptionSender.setReplyToAddress(replyToAddress);
-		subscriptionSender.setScopeGroupId(groupId);
-		subscriptionSender.setServiceContext(serviceContext);
-		subscriptionSender.setUniqueMailId(false);
-
-		return subscriptionSender;
+		return null;
 	}
 
+	/**
+	 * @deprecated As of 7.0.0, with no direct replacement.
+	 */
+	@Deprecated
 	protected void notifyDiscussionSubscribers(
 			long userId, MBMessage message, ServiceContext serviceContext)
 		throws PortalException {
-
-		if (!PrefsPropsUtil.getBoolean(
-				message.getCompanyId(),
-				PropsKeys.DISCUSSION_EMAIL_COMMENTS_ADDED_ENABLED)) {
-
-			return;
-		}
-
-		MBDiscussion mbDiscussion =
-			mbDiscussionLocalService.getThreadDiscussion(message.getThreadId());
-
-		String contentURL = (String)serviceContext.getAttribute("contentURL");
-
-		contentURL = HttpUtil.addParameter(
-			contentURL, serviceContext.getAttribute("namespace") + "messageId",
-			message.getMessageId());
-
-		String userAddress = StringPool.BLANK;
-		String userName = (String)serviceContext.getAttribute(
-			"pingbackUserName");
-
-		if (Validator.isNull(userName)) {
-			userAddress = PortalUtil.getUserEmailAddress(message.getUserId());
-			userName = PortalUtil.getUserName(
-				message.getUserId(), StringPool.BLANK);
-		}
-
-		String fromName = PrefsPropsUtil.getString(
-			message.getCompanyId(), PropsKeys.ADMIN_EMAIL_FROM_NAME);
-		String fromAddress = PrefsPropsUtil.getString(
-			message.getCompanyId(), PropsKeys.ADMIN_EMAIL_FROM_ADDRESS);
-
-		String subject = PrefsPropsUtil.getContent(
-			message.getCompanyId(), PropsKeys.DISCUSSION_EMAIL_SUBJECT);
-		String body = PrefsPropsUtil.getContent(
-			message.getCompanyId(), PropsKeys.DISCUSSION_EMAIL_BODY);
-
-		SubscriptionSender subscriptionSender = new SubscriptionSender();
-
-		subscriptionSender.setBody(body);
-		subscriptionSender.setCompanyId(message.getCompanyId());
-		subscriptionSender.setClassName(MBDiscussion.class.getName());
-		subscriptionSender.setClassPK(mbDiscussion.getDiscussionId());
-		subscriptionSender.setContextAttribute(
-			"[$COMMENTS_BODY$]", message.getBody(message.isFormatBBCode()),
-			false);
-		subscriptionSender.setContextAttributes(
-			"[$COMMENTS_USER_ADDRESS$]", userAddress, "[$COMMENTS_USER_NAME$]",
-			userName, "[$CONTENT_URL$]", contentURL);
-		subscriptionSender.setCurrentUserId(userId);
-		subscriptionSender.setEntryTitle(message.getBody());
-		subscriptionSender.setEntryURL(contentURL);
-		subscriptionSender.setFrom(fromAddress, fromName);
-		subscriptionSender.setHtmlFormat(true);
-
-		Date modifiedDate = message.getModifiedDate();
-
-		subscriptionSender.setMailId(
-			"mb_discussion", message.getCategoryId(), message.getMessageId(),
-			modifiedDate.getTime());
-
-		int notificationType =
-			UserNotificationDefinition.NOTIFICATION_TYPE_ADD_ENTRY;
-
-		if (serviceContext.isCommandUpdate()) {
-			notificationType =
-				UserNotificationDefinition.NOTIFICATION_TYPE_UPDATE_ENTRY;
-		}
-
-		subscriptionSender.setNotificationType(notificationType);
-
-		String portletId = PortletProviderUtil.getPortletId(
-			Comment.class.getName(), PortletProvider.Action.VIEW);
-
-		subscriptionSender.setPortletId(portletId);
-
-		subscriptionSender.setScopeGroupId(message.getGroupId());
-		subscriptionSender.setServiceContext(serviceContext);
-		subscriptionSender.setSubject(subject);
-		subscriptionSender.setUniqueMailId(false);
-
-		String className = (String)serviceContext.getAttribute("className");
-		long classPK = ParamUtil.getLong(serviceContext, "classPK");
-
-		subscriptionSender.addPersistedSubscribers(className, classPK);
-
-		subscriptionSender.flushNotificationsAsync();
 	}
 
+	/**
+	 * @deprecated As of 7.0.0, with no direct replacement.
+	 */
+	@Deprecated
 	protected void notifySubscribers(
 			long userId, MBMessage message, String messageURL,
 			ServiceContext serviceContext)
 		throws PortalException {
-
-		if (!message.isApproved() || Validator.isNull(messageURL)) {
-			return;
-		}
-
-		if (message.isDiscussion()) {
-			try {
-				notifyDiscussionSubscribers(userId, message, serviceContext);
-			}
-			catch (Exception e) {
-				_log.error(e, e);
-			}
-
-			return;
-		}
-
-		MBGroupServiceSettings mbGroupServiceSettings =
-			MBGroupServiceSettings.getInstance(message.getGroupId());
-
-		if (serviceContext.isCommandAdd() &&
-			mbGroupServiceSettings.isEmailMessageAddedEnabled()) {
-		}
-		else if (serviceContext.isCommandUpdate() &&
-				 mbGroupServiceSettings.isEmailMessageUpdatedEnabled()) {
-		}
-		else {
-			return;
-		}
-
-		Company company = companyPersistence.findByPrimaryKey(
-			message.getCompanyId());
-
-		User user = userPersistence.findByPrimaryKey(userId);
-
-		String emailAddress = user.getEmailAddress();
-		String fullName = user.getFullName();
-
-		if (message.isAnonymous()) {
-			emailAddress = StringPool.BLANK;
-			fullName = serviceContext.translate("anonymous");
-		}
-
-		MBCategory category = message.getCategory();
-
-		List<Long> categoryIds = new ArrayList<>();
-
-		categoryIds.add(message.getCategoryId());
-
-		if (message.getCategoryId() !=
-				MBCategoryConstants.DEFAULT_PARENT_CATEGORY_ID) {
-
-			categoryIds.addAll(category.getAncestorCategoryIds());
-		}
-
-		String entryTitle = message.getSubject();
-
-		String fromName = mbGroupServiceSettings.getEmailFromName();
-		String fromAddress = mbGroupServiceSettings.getEmailFromAddress();
-
-		String replyToAddress = StringPool.BLANK;
-
-		if (PropsValues.POP_SERVER_NOTIFICATIONS_ENABLED) {
-			replyToAddress = MBUtil.getReplyToAddress(
-				message.getCategoryId(), message.getMessageId(),
-				company.getMx(), fromAddress);
-		}
-
-		LocalizedValuesMap subjectLocalizedValuesMap = null;
-		LocalizedValuesMap bodyLocalizedValuesMap = null;
-
-		if (serviceContext.isCommandUpdate()) {
-			subjectLocalizedValuesMap =
-				mbGroupServiceSettings.getEmailMessageUpdatedSubject();
-			bodyLocalizedValuesMap =
-				mbGroupServiceSettings.getEmailMessageUpdatedBody();
-		}
-		else {
-			subjectLocalizedValuesMap =
-				mbGroupServiceSettings.getEmailMessageAddedSubject();
-			bodyLocalizedValuesMap =
-				mbGroupServiceSettings.getEmailMessageAddedBody();
-		}
-
-		boolean htmlFormat = mbGroupServiceSettings.isEmailHtmlFormat();
-
-		String messageBody = message.getBody();
-
-		if (htmlFormat && message.isFormatBBCode()) {
-			try {
-				messageBody = BBCodeTranslatorUtil.getHTML(messageBody);
-
-				HttpServletRequest request = serviceContext.getRequest();
-
-				if (request != null) {
-					ThemeDisplay themeDisplay =
-						(ThemeDisplay)request.getAttribute(
-							WebKeys.THEME_DISPLAY);
-
-					messageBody = MBUtil.replaceMessageBodyPaths(
-						themeDisplay, messageBody);
-				}
-			}
-			catch (Exception e) {
-				_log.error(
-					"Unable to parse message " + message.getMessageId() + ": " +
-						e.getMessage());
-			}
-		}
-
-		String inReplyTo = null;
-		String messageSubject = message.getSubject();
-		String messageSubjectPrefix = StringPool.BLANK;
-
-		if (message.getParentMessageId() !=
-				MBMessageConstants.DEFAULT_PARENT_MESSAGE_ID) {
-
-			MBMessage parentMessage = mbMessageLocalService.getMessage(
-				message.getParentMessageId());
-
-			Date modifiedDate = parentMessage.getModifiedDate();
-
-			inReplyTo = PortalUtil.getMailId(
-				company.getMx(), MBUtil.MESSAGE_POP_PORTLET_PREFIX,
-				message.getCategoryId(), parentMessage.getMessageId(),
-				modifiedDate.getTime());
-
-			if (messageSubject.startsWith(
-					MBMessageConstants.MESSAGE_SUBJECT_PREFIX_RE)) {
-
-				messageSubjectPrefix =
-					MBMessageConstants.MESSAGE_SUBJECT_PREFIX_RE;
-
-				messageSubject = messageSubject.substring(
-					messageSubjectPrefix.length());
-			}
-		}
-
-		SubscriptionSender subscriptionSender = getSubscriptionSender(
-			userId, category, message, messageURL, entryTitle, htmlFormat,
-			messageBody, messageSubject, messageSubjectPrefix, inReplyTo,
-			fromName, fromAddress, replyToAddress, emailAddress, fullName,
-			subjectLocalizedValuesMap, bodyLocalizedValuesMap, serviceContext);
-
-		subscriptionSender.addPersistedSubscribers(
-			MBCategory.class.getName(), message.getGroupId());
-
-		for (long categoryId : categoryIds) {
-			if (categoryId != MBCategoryConstants.DEFAULT_PARENT_CATEGORY_ID) {
-				subscriptionSender.addPersistedSubscribers(
-					MBCategory.class.getName(), categoryId);
-			}
-		}
-
-		subscriptionSender.addPersistedSubscribers(
-			MBThread.class.getName(), message.getThreadId());
-
-		subscriptionSender.flushNotificationsAsync();
-
-		if (!MailingListThreadLocal.isSourceMailingList()) {
-			for (long categoryId : categoryIds) {
-				MBSubscriptionSender sourceMailingListSubscriptionSender =
-					getSubscriptionSender(
-						userId, category, message, messageURL, entryTitle,
-						htmlFormat, messageBody, messageSubject,
-						messageSubjectPrefix, inReplyTo, fromName, fromAddress,
-						replyToAddress, emailAddress, fullName,
-						subjectLocalizedValuesMap, bodyLocalizedValuesMap,
-						serviceContext);
-
-				sourceMailingListSubscriptionSender.setBulk(false);
-
-				sourceMailingListSubscriptionSender.addMailingListSubscriber(
-					message.getGroupId(), categoryId);
-
-				sourceMailingListSubscriptionSender.flushNotificationsAsync();
-			}
-		}
 	}
 
 	protected void pingPingback(
@@ -2630,21 +2263,6 @@ public class MBMessageLocalServiceImpl extends MBMessageLocalServiceBaseImpl {
 			int max = PropsValues.DISCUSSION_MAX_COMMENTS - 1;
 
 			throw new DiscussionMaxCommentsException(count + " exceeds " + max);
-		}
-	}
-
-	private String _getLocalizedRootCategoryName(long groupId, Locale locale) {
-		try {
-			Group group = groupPersistence.findByPrimaryKey(groupId);
-
-			return LanguageUtil.get(locale, "message-boards-home") + " - " +
-				group.getDescriptiveName(locale);
-		}
-		catch (PortalException pe) {
-			_log.error(
-				"Unable to get descriptive name for group " + groupId, pe);
-
-			return LanguageUtil.get(locale, "message-boards-home");
 		}
 	}
